+++ beam-core/src/main/java/org/esa/beam/dataio/dimap/DimapProductHelpers.java
@@ -1343,7 +1343,7 @@ public class DimapProductHelpers {
                             gridInfo.getChildTextTrim(DimapProductConstants.TAG_TIE_POINT_STEP_X));
                     final float subsY = Float.parseFloat(
                             gridInfo.getChildTextTrim(DimapProductConstants.TAG_TIE_POINT_STEP_Y));
-                    final float[] floats = new float[width * height];
+                    final double[] floats = new double[width * height];
                     boolean cyclic = false;
                     final String cyclicText = gridInfo.getChildTextTrim(DimapProductConstants.TAG_TIE_POINT_CYCLIC);
                     if (cyclicText != null) {
